recursion: fix compiler warning in guest
add `#[cfg(not(target_os = "zkvm"))]` to items that are not needed for guest compilation
