+++ risc0/zkvm/src/recursion/circuit_impl.rs
@@ -1,0 +1,36 @@
+// Copyright 2023 RISC Zero, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use crate::recursion::TAPSET;
+use risc0_zkp::taps::TapSet;
+use crate::recursion::TapsProvider;
+use crate::recursion::CircuitCoreDef;
+use risc0_zkp::field::baby_bear::BabyBear;
+use crate::recursion::CircuitImpl;
+
+impl CircuitImpl {
+    const fn new() -> Self {
+        CircuitImpl
+    }
+}
+
+impl TapsProvider for CircuitImpl {
+    fn get_taps(&self) -> &'static TapSet<'static> {
+        TAPSET
+    }
+}
+
+impl CircuitCoreDef<BabyBear> for CircuitImpl {}
+
+pub(crate) const CIRCUIT_CORE: CircuitImpl = CircuitImpl::new();
